fix: correctly use config flag by making it a map as it should have been all along
